+++ examples/mdx-defaults-from-remark/src/pages/index.mdx
@@ -1,7 +1,7 @@
 # Hi people
 
 Welcome to your new Gatsby site.
-Now go build something great.
+Now go build something great with `gatsby develop` or other fun commands.
 
 ```js
 const some = {}
