--- packages/gatsby-component-prismjs/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@reux/gatsby-component-prismjs",
-  "version": "0.0.1",
-  "main": "index.js",
-  "author": "Chris Biscardi <chris@christopherbiscardi.com> (@chrisbiscardi)",
-  "license": "MIT",
-  "dependencies": {
-    "mdx-utils": "^0.0.2",
-    "prism-react-renderer": "^0.1.5"
-  }
-}
