+++ packages/gatsby-mdx/package.json
@@ -17,6 +17,7 @@
     "escape-string-regexp": "^1.0.5",
     "fs-extra": "^7.0.0",
     "gray-matter": "^4.0.1",
+    "loader-utils": "^1.2.3",
     "lodash": "^4.17.10",
     "mdast-util-to-string": "^1.0.4",
     "mdast-util-toc": "^2.0.1",
