+++ packages/gatsby-plugin-mdx-prismjs/index.js
@@ -1,8 +1,10 @@
-const React = require("react");
+const {
+  GatsbyComponentPrismjsReplacement,
+  GatsbyComponentPrismjsReplacementInline
+} = require("./components");
 
 exports.guards = {};
 exports.components = {
-  inlineCode: props => React.createElement("code", undefined, props),
-  pre: props =>
-    console.log("prism props", props) || React.createElement("pre", props)
+  inlineCode: GatsbyComponentPrismjsReplacementInline,
+  pre: GatsbyComponentPrismjsReplacement
 };
